requirements.txt
streamlit>=1.22.0
requests>=2.28.1
beautifulsoup4>=4.11.1
ollama>=0.1.5
python-dotenv>=1.0.0